AArch64: Fixed some lowering issues.
Before there was not a proper separation of between AMD64 and AArch64 specific
lowerings.

Refactored architecture-specific functionality into separate classes.
Also added proper AArch64 RemNode lowering support, and removed
AMD64-specific FloatConvertNode and AMD64ArrayIndexOfDispatchNode
lowering from AArch64 path.
