+++ substratevm/src/com.oracle.svm.core.graal/src/com/oracle/svm/core/graal/meta/SubstrateBasicLoweringProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2017, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -45,9 +45,7 @@ import org.graalvm.compiler.nodes.NodeView;
 import org.graalvm.compiler.nodes.StructuredGraph;
 import org.graalvm.compiler.nodes.ValueNode;
 import org.graalvm.compiler.nodes.calc.AndNode;
-import org.graalvm.compiler.nodes.calc.RemNode;
 import org.graalvm.compiler.nodes.calc.UnsignedRightShiftNode;
-import org.graalvm.compiler.nodes.extended.ForeignCallNode;
 import org.graalvm.compiler.nodes.extended.LoadHubNode;
 import org.graalvm.compiler.nodes.java.NewArrayNode;
 import org.graalvm.compiler.nodes.java.NewInstanceNode;
@@ -63,7 +61,6 @@ import org.graalvm.compiler.options.OptionValues;
 import org.graalvm.compiler.phases.util.Providers;
 import org.graalvm.compiler.replacements.DefaultJavaLoweringProvider;
 import org.graalvm.compiler.replacements.SnippetCounter.Group;
-import org.graalvm.compiler.replacements.amd64.AMD64ArrayIndexOfDispatchNode;
 import org.graalvm.compiler.replacements.nodes.AssertionNode;
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
@@ -130,19 +127,10 @@ public abstract class SubstrateBasicLoweringProvider extends DefaultJavaLowering
         return lowerings;
     }
 
-    @SuppressWarnings("unchecked")
     @Override
     public void lower(Node n, LoweringTool tool) {
-        @SuppressWarnings("rawtypes")
-        NodeLoweringProvider lowering = lowerings.get(n.getClass());
-        if (lowering != null) {
-            lowering.lower(n, tool);
-        } else if (n instanceof RemNode) {
-            /* No lowering necessary. */
-        } else if (n instanceof AssertionNode) {
+        if (n instanceof AssertionNode) {
             lowerAssertionNode((AssertionNode) n);
-        } else if (n instanceof AMD64ArrayIndexOfDispatchNode) {
-            lowerArrayIndexOf((AMD64ArrayIndexOfDispatchNode) n);
         } else {
             super.lower(n, tool);
         }
@@ -237,12 +225,6 @@ public abstract class SubstrateBasicLoweringProvider extends DefaultJavaLowering
     private static void lowerAssertionNode(AssertionNode n) {
         // we discard the assertion if it was not handled by any other lowering
         n.graph().removeFixed(n);
-    }
-
-    private void lowerArrayIndexOf(AMD64ArrayIndexOfDispatchNode dispatchNode) {
-        StructuredGraph graph = dispatchNode.graph();
-        ForeignCallNode call = graph.add(new ForeignCallNode(foreignCalls, dispatchNode.getStubCallDescriptor(), dispatchNode.getStubCallArgs()));
-        graph.replaceFixed(dispatchNode, call);
     }
 
     @Override
