+++ substratevm/src/com.oracle.svm.core.graal/src/com/oracle/svm/core/graal/snippets/ArithmeticSnippets.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2017, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -61,7 +61,7 @@ import jdk.vm.ci.meta.DeoptimizationAction;
 import jdk.vm.ci.meta.DeoptimizationReason;
 import jdk.vm.ci.meta.JavaKind;
 
-public final class ArithmeticSnippets extends SubstrateTemplates implements Snippets {
+public abstract class ArithmeticSnippets extends SubstrateTemplates implements Snippets {
 
     @Snippet
     protected static int idivSnippet(int x, int y, @ConstantParameter boolean needsZeroCheck) {
@@ -179,12 +179,6 @@ public final class ArithmeticSnippets extends SubstrateTemplates implements Snip
 
     private final ObjectLayout layout;
 
-    @SuppressWarnings("unused")
-    public static void registerLowerings(OptionValues options, Iterable<DebugHandlersFactory> factories, Providers providers, SnippetReflectionProvider snippetReflection,
-                    Map<Class<? extends Node>, NodeLoweringProvider<?>> lowerings) {
-        new ArithmeticSnippets(options, factories, providers, snippetReflection, lowerings);
-    }
-
     private final SnippetInfo idiv;
     private final SnippetInfo ldiv;
     private final SnippetInfo irem;
@@ -194,7 +188,7 @@ public final class ArithmeticSnippets extends SubstrateTemplates implements Snip
     private final SnippetInfo uirem;
     private final SnippetInfo ulrem;
 
-    private ArithmeticSnippets(OptionValues options, Iterable<DebugHandlersFactory> factories, Providers providers, SnippetReflectionProvider snippetReflection,
+    protected ArithmeticSnippets(OptionValues options, Iterable<DebugHandlersFactory> factories, Providers providers, SnippetReflectionProvider snippetReflection,
                     Map<Class<? extends Node>, NodeLoweringProvider<?>> lowerings) {
         super(options, factories, providers, snippetReflection);
         this.layout = ConfigurationValues.getObjectLayout();
@@ -245,7 +239,7 @@ public final class ArithmeticSnippets extends SubstrateTemplates implements Snip
         }
     }
 
-    static class IdentityLowering implements NodeLoweringProvider<Node> {
+    public static class IdentityLowering implements NodeLoweringProvider<Node> {
         @Override
         public void lower(Node node, LoweringTool tool) {
             // do nothing and leave node unchanged.
